refactor & fix: update expression precedence mechanism
This patch teaches binary expressions to be correctly grouped
using a stack of previous unterminated expressions before current one
and binary operators in between.
The stack reduces (folds) expressions when encountered a binary operator
which has precedence lower than currently highest precedence, however,
only binary operators which have precendence at least as the encountered
operator are reduced. This way we wait to group expressions until it is
safe to do so.

There are a couple of new FIXMEs, which will be adressed in commits to
follow!