feat: implement printf function
This patch adds support for the same abilities as in the print statement
as a function.

Parser was not liking printf function because it was defined as a keyword.

Also I noticed that tests failed intermittently, probably due to
tests running in parallel. Thus changing filenames in tests eliminated
this problem because processes did not have to contend for the same files.